pyproject.toml
[project]
name = "thatrack-bot"
version = "1.0.0"
description = "a telegram bot for thatrack"
authors = [
    {name = "thediogen",email = "evgendvodnenko@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "aiogram (>=3.20.0.post0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "aioredis (>=2.0.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
